fix requires of input string format, replaced double to int

+++ src/Calculator.java
@@ -1,7 +1,7 @@
 public class Calculator {
 
-        public static double calculate ( double var1, double var2, char operation){
-            double result = 0;
+        public static int calculate ( int var1, int var2, char operation){
+            int result = 0;
             switch (operation) {
 
                 case '+':
